Moved broadcasts to separate class
